Update .gitignore
Addition of /vocab route to flask app. New postgresql non spatial provider.

Update to plugin name

Update to collection to add in license information if present.

Update to collection to add in license information if present.

Update to collection to add in license information if present.

Update to collection to add in license information if present.

Update to collection to add in license information if present.

Update to collection to add in license information if present.

Update to collection to add in license information if present.

Update to collection to add in license information if present.

Update to collection to add in license information if present.

CRUD operations for psql.

CRUD operations for psql.

CRUD operations for psql.

CRUD operations for psql.

shapely 2

shapely 2

debugging

debugging

debugging

debugging

debugging

debugging

debugging

sqlalchemy < 2 !

testing

testing

addition of delete

addition of delete